+++ mapmatching/utils/timer.py
@@ -36,8 +36,8 @@ def timeit(func):
         res = func(*args, **kwargs)
         end = time.time()
         _log = f"{func.__name__}, cost: {end - start: .3f} s"
-        print(_log)
-        # logger.info(_log)
+        # print(_log)
+        logger.info(_log)
         
         return res